feat(api) rename stl-api-gen to .stl-codegen, support more module resolution types
