+++ demo/.stl-codegen/api/posts/create.ts
