+++ demo/.stl-codegen/api/posts/models.ts
