+++ demo/.stl-codegen/api/posts/retrieve.ts
