+++ demo/libs/stl.ts
@@ -5,7 +5,7 @@ import { makeNextAuthPlugin } from "@stl-api/next-auth";
 import { authOptions } from "../pages/api/auth/[...nextauth]";
 import { makeCurrentUserPlugin } from "./currentUserPlugin";
 import { User } from "@prisma/client";
-import { typeSchemas } from "../stl-api-gen";
+import { typeSchemas } from "../.stl-codegen";
 
 declare module "stainless" {
   interface StlCustomContext {
@@ -23,5 +23,5 @@ const plugins = {
 
 export const stl = new Stl({
   plugins,
-  typeSchemas
+  typeSchemas,
 });
