--- demo/stl-api-gen/index.ts
@@ -1,6 +1,0 @@
-export const typeSchemas = {
-  "post /api/posts": () =>
-    import("./api/posts/create").then((mod) => mod.post__api_posts),
-  "get /api/posts/{post}": () =>
-    import("./api/posts/retrieve").then((mod) => mod.get__api_posts_$post$),
-};
